requirements.txt
aiofiles==22.1.0
aiosqlite==0.18.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.0.5
attrs==22.2.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
certifi==2022.9.24
cffi==1.15.1
chain==1.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.4
colorclass==2.2.2
comm==0.1.2
compiler==0.2.0
contourpy==1.0.5
cvxopt==1.3.0
cvxpy==1.2.1
cycler==0.11.0
Cython==0.29.32
dash==2.8.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
docopt==0.6.2
ecos==2.0.10
git+ssh://git@bitbucket.org/enscodevelopment/enscohelper.git@datafeed2
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fastjsonschema==2.16.2
Flask==2.2.2
fonttools==4.37.4
fqdn==1.5.1
future==0.18.2
greenlet==1.1.3
idna==3.4
importlib-metadata==6.0.0
ipykernel==6.14
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.5.0
jupyter-events==0.6.3
jupyter-ydoc==0.2.2
jupyter_client==7.4.9
jupyter_core==5.2.0
jupyter_server==2.2.1
jupyter_server_fileid==0.7.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.6.1
jupyterlab==3.6.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
jupyterlab_server==2.19.0
kiwisolver==1.4.4
lxml==4.9.1
MarkupSafe==2.1.2
matplotlib==3.6.1
matplotlib-inline==0.1.3
mistune==2.0.5
mpmath==1.2.1
mysql-connector-python==8.0.30
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.4
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.3
openpyxl==3.0.10
osqp==0.6.2.post5
packaging==21.3
pandas==1.5.0
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
pickleshare==0.7.5
Pillow==9.4.0
pip-upgrader==1.4.15
platformdirs==3.0.0
plotly==5.13.0
ply==3.11
pmdarima==2.0.1
prometheus-client==0.16.0
prompt-toolkit==3.0.30
protobuf==3.20.1
psutil==5.9.0
psycopg2-binary==2.9.3
pure-eval==0.2.2
pycparser==2.21
pygame==2.1.3
Pygments==2.11.2
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.5
pytz==2022.2.1
pywin32==303
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.0
qdldl==0.1.5.post2
qtconsole==5.4.0
QtPy==2.3.0
requests==2.28.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scikit-learn==1.1.2
scipy==1.9.3
scs==3.2.0
Send2Trash==1.8.0
sentry-sdk==1.9.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
sseclient-py==1.7.2
stack-data==0.2.0
statsmodels==0.13.2
sympy==1.11.1
tenacity==8.1.0
terminado==0.17.1
terminaltables==3.1.10
threadpoolctl==3.1.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
traitlets==5.9.0
uri-template==1.2.0
urllib3==1.26.12
wapi-python==0.7.7
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.2
widgetsnbextension==4.0.5
XlsxWriter==3.0.3
y-py==0.5.9
ypy-websocket==0.8.2
zipp==3.12.0
